chatgpt.py
#pip install openai boto3 wit
#pip install azure-cognitiveservices-knowledge-qnamaker msrest
#pip install dialogflow google-cloud-dialogflow
#pip install ibm-watson ibm-cloud-sdk-core

#OpenAI imports
import openai
import os
#Google Lambda imports
import boto3
#Wit.ai imports
from wit import Wit
#Microsoft Azure imports
from azure.cognitiveservices.knowledge.qnamaker import QnAMakerClient
from msrest.authentication import CognitiveServicesCredentials
#DialogFlow imports
import dialogflow_v2 as dialogflow
from google.api_core.exceptions import InvalidArgument
#IBM Watson imports
import json
from ibm_watson import AssistantV2
from ibm_cloud_sdk_core.authenticators import IAMAuthenticator



# Prompt the user for a question
question = input("Ask a question: ")



# Set up the OpenAI API client
openai.api_key = os.environ["OPENAI_API_KEY"]
# Send the question to the OpenAI API
response = openai.Completion.create(
    engine="davinci",
    prompt=f"Q: {question}\nA:",
    temperature=0.5,
    max_tokens=1024,
    n=1,
    stop=None,
)
# Extract the answer from the API response
answer = response.choices[0].text.strip()
# Display the answer
print(answer)



# Create a Lambda client
lambda_client = boto3.client('lambda')
# Invoke the Lambda function with the question as the payload
response = lambda_client.invoke(
    FunctionName='my-lambda-function',
    Payload=f'{{"question": "{question}"}}'
)
# Extract the answer from the Lambda response
answer = response['Payload'].read().decode()
# Display the answer
print(answer)



# Set up the Wit client
access_token = 'YOUR_ACCESS_TOKEN'
client = Wit(access_token=access_token)
# Send the question to Wit.ai
response = client.message(question)
# Extract the answer from the Wit.ai response
answer = response['intents'][0]['name']
# Display the answer
print(answer)



# Create a QnA Maker client
qna_maker_endpoint = "https://YOUR_QNA_MAKER_ENDPOINT_NAME.azurewebsites.net"
qna_maker_subscription_key = "YOUR_QNA_MAKER_SUBSCRIPTION_KEY"
qna_maker_client = QnAMakerClient(qna_maker_endpoint, CognitiveServicesCredentials(qna_maker_subscription_key))
# Send the question to the QnA Maker service
response = qna_maker_client.qna.generate_answer("YOUR_KNOWLEDGE_BASE_ID", {"question": question})
# Extract the answer from the QnA Maker response
answer = response.answers[0].answer
# Display the answer
print(answer)



# Set up the Dialogflow client
project_id = 'YOUR_PROJECT_ID'
session_id = 'YOUR_SESSION_ID'
language_code = 'en-US'
session_client = dialogflow.SessionsClient()
session = session_client.session_path(project_id, session_id)
# Send the question to Dialogflow
text_input = dialogflow.types.TextInput(text=question, language_code=language_code)
query_input = dialogflow.types.QueryInput(text=text_input)
try:
    response = session_client.detect_intent(session=session, query_input=query_input)
except InvalidArgument:
    raise
# Extract the answer from the Dialogflow response
answer = response.query_result.fulfillment_text
# Display the answer
print(answer)



# Set up the Watson Assistant client
authenticator = IAMAuthenticator('YOUR_API_KEY')
assistant = AssistantV2(
    version='2019-02-28',
    authenticator=authenticator
)
assistant.set_service_url('YOUR_SERVICE_URL')
assistant_id = 'YOUR_ASSISTANT_ID'
# Send the question to Watson
response = assistant.message(
    assistant_id=assistant_id,
    session_id='unique',
    input={
        'message_type': 'text',
        'text': question
    }
).get_result()
# Extract the answer from the Watson response
answer = response['output']['generic'][0]['text']
# Display the answer
print(answer)
