requirements.txt
# pip3 install -r requirements.txt

# core

# packaging

wheel
pip
setuptools
packaging

# HTML XML JSON YAML

urllib3
requests
oauthlib
requests-oauthlib
google-auth
google-auth-oauthlib
webencodings
pyyaml
confuse
soupsieve
beautifulsoup4
htmlmin
markdown
MarkupSafe
jinja2
jsonschema
tornado
parso
jedi
pygments
mistune
pyparsing
bleach
pandocfilters
defusedxml
et-xmlfile
werkzeug

# windows

windows-ncurses
pywin32
pywinpty

# gaming

pyopengl
pygame
freegames
pymunk

# audio

forex-python 
scapy
wave
mutagen
pyaudio

# encryption

pynacl
cryptography
rsa
bcrypt
pgpy
imagehash

# speech

translate
googletrans
gtts-token
gtts
pyttsx
SpeechRecognition

# timezone dates

pytz
dateutils
python-dateutil
jdcal

# Python 2

six

# database

mongoengine
pymongo
psycopg2
pickleshare

# graphics

cycler
local
matplotlib
matplotlib-venn
Pillow
cairosvg
seaborn
missingno

# unknown

confuse
backcall
pandocfilters

# interface

tqdm
colorama
click
termcolor

# UniCode

tangled-up-in-unicode
wcwidth
prompt-toolkit
chardet
idna

# jupyter notebook

notebook
ipykernel
ipython-genutils
ipython
jupyter-core
jupyter-client
nbformat
nbconvert
terminado
ipywidgets
widgetsnbextension
pivottablejs

# enhancements

certifi
cachetools
decorator
traitlets
attrs
joblib
entrypoints
testpath
Send2Trash
kiwisolver
wrapt

# system messaging queue

pyzmq
protobuf
pyasn1
pyasn1-modules
grpcio

# data-type processing

numpy
scipy
mpmath
sympy
astropy
astunparse
pandas
llvmlite
numba
gast
pyreadstat 
prometheus-client
pyrsistent
PyWavelets
openpyxl
xlsxwriter
h5py

# machine learning

networkx
visions
phik
scikit-learn
absl-py
tensorboard-plugin-wit
tensorboard
tensorflow-estimator
tensorflow
tensorflow-gpu
opencv-python
openai
openai-gym
openai-research
openai-baselines
opt-einsum
keras-preprocessing
pandas-profiling
google-pasta

# multi-threading

threadpoolctl

# wheel, pyparsing, six, packaging, urllib3, chardet, idna, certifi, requests, oauthlib, requests-oauthlib, pyasn1, pyasn1-modules, cachetools, rsa, google-auth, google-auth-oauthlib, webencodings, pyyaml, confuse, soupsieve, beautifulsoup4, htmlmin, markdown, MarkupSafe, jinja2, pyrsistent, attrs, jsonschema, tornado, parso, jedi, pygments, mistune, bleach, pandocfilters, defusedxml, et-xmlfile, werkzeug, pywin32, pywinpty, pygame, wave, mutagen, pyaudio, pycparser, cffi, bcrypt, numpy, scipy, Pillow, pyreadstat, PyWavelets, imagehash, gtts-token, click, gtts, pyttsx, pytz, python-dateutil, jdcal, pymongo, mongoengine, psycopg2, pickleshare, cycler, kiwisolver, matplotlib, matplotlib-venn, cairocffi, tinycss2, cssselect2, cairosvg, pandas, seaborn, missingno, backcall, tqdm, colorama, termcolor, tangled-up-in-unicode, wcwidth, prompt-toolkit, ipython-genutils, decorator, traitlets, jupyter-core, pyzmq, jupyter-client, prometheus-client, Send2Trash, nbformat, ipython, ipykernel, terminado, testpath, entrypoints, nbconvert, notebook, widgetsnbextension, ipywidgets, pivottablejs, joblib, wrapt, protobuf, grpcio, mpmath, sympy, astunparse, llvmlite, numba, gast, openpyxl, xlsxwriter, h5py, networkx, visions, phik, threadpoolctl, scikit-learn, absl-py, tensorboard-plugin-wit, tensorboard, tensorflow-estimator, keras-preprocessing, opt-einsum, google-pasta, tensorflow, opencv-python, openai, openai-gym, openai-research, openai-baselines